i think im caught up? feel like im missing something

+++ ResilienceGame/Assets/Cards/CardData.cs
@@ -23,9 +23,9 @@ public struct CardData
     public float percentSuccess;
     public int cardID;
     public int teamID; // TODO: Use for SD
-    public int blueCost;
-    public int blackCost;
-    public int purpleCost;
+    public float blueCost;
+    public float blackCost;
+    public float purpleCost;
     public int drawAmount;
     public int removeAmount;
     public int targetAmount;
